requirements.txt

numpy==2.0.0
opencv_contrib_python==4.9.0.80
Pillow==10.4.0
asgiref==3.7.2
Django==5.0.2
numpy==1.26.4
psycopg==3.1.18
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytorch-transformers==1.2.0
sqlparse==0.4.4
torch==2.2.1
transformers==4.38.2
tzdata==2023.4
wheel==0.42.0
whitenoise==6.7.0